requirements.txt
kafka-python-ng==2.2.2
websocket-client==1.8.0
python-dotenv==1.0.1